+++ src/risedevtool/src/task.rs
@@ -40,15 +40,13 @@ mod tempo_service;
 mod utils;
 
 use std::env;
-use std::io::{Read as _, Seek as _, SeekFrom};
 use std::net::{TcpStream, ToSocketAddrs};
 use std::path::{Path, PathBuf};
 use std::process::{Command, Output};
 use std::sync::Arc;
 use std::time::Duration;
 
-use anyhow::{anyhow, bail, Context, Result};
-use fs_err::File;
+use anyhow::{anyhow, Context, Result};
 use indicatif::ProgressBar;
 use reqwest::blocking::{Client, Response};
 use tempfile::TempDir;
@@ -188,36 +186,6 @@ where
 
     pub fn log_path(&self) -> &Path {
         self.log_file.as_ref().unwrap().as_path()
-    }
-
-    pub fn wait_log_contains(&mut self, pattern: impl AsRef<str>) -> anyhow::Result<()> {
-        let pattern = pattern.as_ref();
-        let log_path = self.log_path().to_path_buf();
-
-        let mut content = String::new();
-        let mut offset = 0;
-
-        wait(
-            || {
-                let mut file = File::open(&log_path).context("log file does not exist")?;
-                file.seek(SeekFrom::Start(offset as u64))?;
-                offset += file.read_to_string(&mut content)?;
-
-                // Always going through the whole log file could be stupid, but it's reliable.
-                if content.contains(pattern) {
-                    Ok(())
-                } else {
-                    bail!("pattern \"{}\" not found in log", pattern)
-                }
-            },
-            &mut self.log,
-            self.status_file.as_ref().unwrap(),
-            self.id.as_ref().unwrap(),
-            Some(Duration::from_secs(30)),
-            true,
-        )?;
-
-        Ok(())
     }
 
     pub fn wait_tcp(&mut self, server: impl AsRef<str>) -> anyhow::Result<()> {
